move attach_server_configuration to clustering_utils
